projetos-gitlab/flask-sql-injection/app/requirements.txt
flask
flask-sqlalchemy
mysqlclient